Gameplay.fxml Scene is built, images need updating
Gameplay.fxml file readied, made sure to have everything properly spaced in case many combos are possible/to make it look decent once things are set to hidden or revealed.
NOTE: I used images that weren't inside the project (in case it's already on the main branch). That needs to be updated with the icons gathered (hopefully already in the project).